# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Haal Lab API",
    "description": "API for interacting with Haal Lab's AI assistant and services",
    "version": "1.0.0",
    "contact": {
      "name": "Haal Lab",
      "email": "hussain.nazary@haal-lab.solutions",
      "url": "https://haal-lab.solutions"
    }
  },
  "servers": [
    {
      "url": "https://haal-lab.solutions",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/chat": {
      "post": {
        "summary": "Chat with Haal Lab AI Assistant",
        "description": "Send a message to the Haal Lab AI assistant to learn about services, capabilities, and get information",
        "operationId": "chatWithAssistant",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["messages"],
                "properties": {
                  "messages": {
                    "type": "array",
                    "description": "Array of chat messages in OpenAI format",
                    "items": {
                      "type": "object",
                      "required": ["role", "content"],
                      "properties": {
                        "role": {
                          "type": "string",
                          "enum": ["user", "assistant", "system"],
                          "description": "The role of the message sender"
                        },
                        "content": {
                          "type": "string",
                          "description": "The message content"
                        }
                      }
                    }
                  }
                }
              },
              "example": {
                "messages": [
                  {
                    "role": "user",
                    "content": "What services does Haal Lab offer?"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "The assistant's response"
                    },
                    "usage": {
                      "type": "object",
                      "description": "Token usage statistics",
                      "properties": {
                        "prompt_tokens": {
                          "type": "integer"
                        },
                        "completion_tokens": {
                          "type": "integer"
                        },
                        "total_tokens": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid message format"
          },
          "401": {
            "description": "Unauthorized - API key issue"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": ["Chat"]
      }
    }
  },
  "components": {
    "schemas": {
      "Message": {
        "type": "object",
        "required": ["role", "content"],
        "properties": {
          "role": {
            "type": "string",
            "enum": ["user", "assistant", "system"]
          },
          "content": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Chat",
      "description": "Interact with the Haal Lab AI assistant"
    }
  ]
}